+++ src/utils/testUtils.tsx
@@ -5,6 +5,7 @@ import {
 } from 'react';
 import {MemoryRouter, Route, Routes} from 'react-router-dom';
 import {I18nextProvider} from 'react-i18next';
+import {Provider} from 'mobx-react';
 import {InitialEntry} from 'history';
 import i18n from '../i18nTestingConfig';
 
@@ -26,7 +27,7 @@ export const renderWithRouter = (
   )
 }
 
-export const renderWithSuspense = (children?: ReactNode | undefined) => (
+export const renderWithSuspense = (children?: ReactNode | undefined): ReactElement => (
   <Suspense fallback='Loading...'>
     {children}
   </Suspense>
@@ -34,6 +35,20 @@ export const renderWithSuspense = (children?: ReactNode | undefined) => (
 
 export const renderWithI18AndSuspenseAndRouter = (
   component: ReactNode, path: string, initialEntries: InitialEntry[]
-) => renderWithI18(renderWithSuspense(renderWithRouter(component, path, initialEntries)))
+): ReactElement => renderWithI18(renderWithSuspense(renderWithRouter(component, path, initialEntries)))
 
-export const renderWithI18AndSuspense = (children?: ReactNode | undefined) => renderWithI18(renderWithSuspense(children))
+export const renderWithI18AndSuspense = (children?: ReactNode | undefined): ReactElement => renderWithI18(renderWithSuspense(children))
+
+export const renderWithMobX = (rootStore: any, children?: ReactNode | undefined): ReactElement => (
+  <Provider rootStore={rootStore}>
+    {children}
+  </Provider>
+)
+
+export const renderWithI18AndSuspenseAndRouterAndMobX = (
+  rootStore: any,
+  component: ReactNode,
+  path: string,
+  initialEntries: InitialEntry[]
+): ReactElement =>
+  renderWithMobX(rootStore, renderWithI18AndSuspenseAndRouter(component, path, initialEntries))