+++ src/views/Characters/components/Characters/Characters.test.tsx
@@ -4,35 +4,10 @@ import {
   screen
 } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import {renderWithI18AndSuspenseAndRouter} from '../../../../utils/testUtils';
-import {RouteName, Vision} from '../../../../const/consts';
+import {renderWithI18AndSuspenseAndRouter, renderWithI18AndSuspenseAndRouterAndMobX} from '../../../../utils/testUtils';
+import {RouteName} from '../../../../const/consts';
 import Characters from './index';
-
-const MOCK_CHARACTERS = [{
-  name: 'Albedo',
-  vision: Vision.GEO,
-  ascension_materials: {},
-  rarity: 5,
-  talent_materials: {},
-  possible_teams: [],
-  attack: {},
-  elemental_skill: {},
-  elemental_burst: {},
-  search_keys: 'albedo альбедо geo гео',
-  translated_name: 'Albedo'
-}, {
-  name: 'Jean',
-  vision: Vision.ANEMO,
-  ascension_materials: {},
-  rarity: 5,
-  talent_materials: {},
-  possible_teams: [],
-  attack: {},
-  elemental_skill: {},
-  elemental_burst: {},
-  search_keys: 'jean джинн anemo анемо',
-  translated_name: 'Jean'
-}]
+import charactersStore from '../../../../store/modules/characters/characters';
 
 describe('Characters', () => {
   const path = RouteName.DEFAULT
@@ -40,7 +15,7 @@ describe('Characters', () => {
 
   test('renders Characters component', async () => {
     render(
-      renderWithI18AndSuspenseAndRouter(<Characters characters={MOCK_CHARACTERS}/>, path, initialEntries)
+      renderWithI18AndSuspenseAndRouter(<Characters/>, path, initialEntries)
     )
 
     expect(screen.getByText(/loading.../i)).toBeInTheDocument()
@@ -51,25 +26,23 @@ describe('Characters', () => {
 
   test('character search works', async () => {
     render(
-      renderWithI18AndSuspenseAndRouter(<Characters characters={MOCK_CHARACTERS}/>, path, initialEntries)
+      renderWithI18AndSuspenseAndRouterAndMobX(charactersStore, <Characters/>, path, initialEntries)
     )
 
-    const charactersBeforeSearch = await screen.findAllByRole('link')
-    expect(charactersBeforeSearch).toHaveLength(2)
-    expect(charactersBeforeSearch[0]).toHaveAttribute('href', '/Jean')
-    expect(charactersBeforeSearch[1]).toHaveAttribute('href', '/Albedo')
+    const charactersBeforeSearch = await screen.findAllByRole('heading', {level: 4});
+    expect(charactersBeforeSearch.length).toBeGreaterThan(50)
 
     const searchField = await screen.findByRole('textbox')
     await waitFor(() => userEvent.type(searchField, 'Albedo'))
 
-    const charactersAfterSearch = await screen.findAllByRole('link')
+    const charactersAfterSearch = await screen.findAllByRole('heading', {level: 4});
     expect(charactersAfterSearch).toHaveLength(1)
-    expect(charactersAfterSearch[0]).toHaveAttribute('href', '/Albedo')
+    expect(charactersAfterSearch[0]).toHaveTextContent('Albedo')
   })
 
   test('onChange function works in search', async () => {
     render(
-      renderWithI18AndSuspenseAndRouter(<Characters characters={MOCK_CHARACTERS}/>, path, initialEntries)
+      renderWithI18AndSuspenseAndRouter(<Characters/>, path, initialEntries)
     )
 
     const searchField = await screen.findByRole('textbox')