+++ src/views/Characters/components/Characters/types.ts
@@ -1,5 +1,0 @@
-import {ICharacterProps as ICharacterWithSearchKeys} from '../Character/types';
-
-export interface ICharactersProps {
-  characters: ICharacterWithSearchKeys[]
-}