+++ src/views/Layout/index.tsx
@@ -1,16 +1,15 @@
 import {FC} from 'react'
 
 import Header from '../Header';
-import CharactersWrapper from '../Characters/components/CharactersWrapper';
 import Container from '../../common/Container';
+import Characters from '../Characters/components/Characters';
 
 const Layout: FC = () => {
   return (
-      <Container>
-        <Header/>
-
-        <CharactersWrapper/>
-      </Container>
+    <Container>
+      <Header/>
+      <Characters/>
+    </Container>
   )
 }